Robot/Parallax
Added a harmless NPC that walks back and forth along a path. He turns
around when there is a cliff edge and when he hits something.

Also added the ability to make any sprite or group of objects parallax
with the camera. A sprite will always parallax to the size of the
camera.